No changes, bulletin #1 (see extended descripton)
The previous commit that mentions including a diff in the commit message
is now required. I am lifting this requirement, as it is superfluous.
GitHub automatically does diffs, as does Git Cola (Linux) and Git Gui
(Windows).

This does not lift the requrement of a sign-off message. In addition to
the other requirements listed, if multiple people worked on a commit it
is required that all of the developers who worked on te code include
their own individual sign-offs. A commit with no sign-off is considered
invalid and will be removed.

Additionally, another requirement has been added that you use Git GUI
for Windows as your tool for making commits and pushing them, or Git
Cola in Linux. I will not accept any commits and/or pull requests from
the GitHub web interface or the GitHub desktop utility. The reasons for
this requirement are similar to the ones expressed by @torvalds for not
accepting PRs through GitHub.

This bulletin goes into effect from this commit onwards.

Signed-off-by: Nkeck720 <noahkeck72@gmail.com>